transition to angular bootstrap
a lot of bootstrap is implemented in jquery. The angular UI team has
created a version of bootstrap that uses angular directives rather than
jquery selectors. I am transitioning the site to use the angular
directives.

This transition is not complete! A dependency injection for angular
bootstrap must be created for this to work!